Move all pipeline parsing code -> pipeline.go
